Rewrite requests to use one common generic method
This removes a lot of redundant code and makes it easier to add new endpoints
by reducing the amount of code that is necessary.
Also all variables that were objects (or arrays) were marked as Readonly so that
the user can be sure that no inserted object (or array) will be manipulated in any
way by this library.